[#148] Refactored Map components
* Added /containers folder in src/
* Added container for each component
* Moved behaviour implementation from component to container
